requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
xlsxwriter>=3.1.0
